+++ src/main/java/lotto/util/constants/LottoConstants.java
@@ -5,7 +5,7 @@ public final class LottoConstants {
     public static final String INPUT_DELIMITER = ",";
     public static final String CONSECUTIVE_DELIMITER = ",,";
 
-    public static final int LOTTO_TICKET_PRICE = 1000;
+    public static final long LOTTO_TICKET_PRICE = 1000;
     public static final int LOTTO_NUMBER_START = 1;
     public static final int LOTTO_NUMBER_END = 45;
     public static final int NEEDED_LOTTO_NUMBER_COUNT = 6;
